feat: implement user recommendation requests
